Renamed the LINZ specific tests and created a benchmark comparison function to reduce repeated code
